==> 07_QMiSeq_MultiTS.R <==
####
#### R code for Ushio et al.
#### "Quantitative monitoring of multispecies fish environmental DNA using high-throughput sequencing"
#### No.6 Illustrations of multispecies fish eDNA time series
####

# names for the output folder and workspace file
qmiseq.out.06 <- "06_QMiSeq_DynamicsOut"
qmiseq.out.07 <- "07_QMiSeq_MultiTSOut"
dir.create(qmiseq.out.07, showWarnings = FALSE)
ws.out.06 <- file.path(qmiseq.out.06, "06_QMiSeq_DynamicsOut.RData")
ws.out.07 <- file.path(qmiseq.out.07, "07_QMiSeq_MultiTSOut.RData")

# load workspace
load(ws.out.06)

# load library
library(ggplot2)
library(cowplot)
library(reshape2)
library(scales)

# select top 10 species
top10.names <- names(sort(colSums(converted.reads.sample, na.rm = T), decreasing = T)[1:10])
top10.copy <- data.frame(date = d.comp$date,
                         na.omit(d.calc.read)[,top10.names])

## normalize eDNA for better appearance
top10.norm <- data.frame(date = d.comp$date,
                         apply(top10.copy[,2:11], 2, function(x) as.numeric(scale(x))))
top10.melt <- melt(top10.norm, id.var="date")

# multispecies time series
# Please note that this is a ggplot version.
# The figure in the main text was generated by Plotly funcutions, which you need your own account if you want to use.
# For more details, please see https://plot.ly/
f6 <- ggplot(top10.melt, aes(x=date, y=value, group=variable, colour=variable))
f6 <- f6 + geom_line(size=0.5)
f6 <- PlotStyle(f6) + ggtitle('Multispecies eDNA time series')
f6 <- f6 + ylab("Normalized eDNA copies")
f6 <- f6 + xlab('Sampling date') + scale_x_datetime(labels=date_format('%Y-%m'))

# output figure
dev.off()
quartz(height=4, width=10)
f6
fig6.name <- file.path(qmiseq.out.00, "Figure6.png")
quartz.save(fig6.name)

# save workspace
save.image(ws.out.07)
